Move all ChefFS exceptions into a single file
This leaves back compat requires to ensure nothing breaks.
