+++ lib/chef/chef_fs/file_system.rb
@@ -17,9 +17,7 @@
 #
 
 require "chef/chef_fs/path_utils"
-require "chef/chef_fs/file_system/default_environment_cannot_be_modified_error"
-require "chef/chef_fs/file_system/operation_failed_error"
-require "chef/chef_fs/file_system/operation_not_allowed_error"
+require "chef/chef_fs/file_system/exceptions"
 require "chef/chef_fs/parallelizer"
 
 class Chef
