+++ lib/chef/chef_fs/file_system/already_exists_error.rb
@@ -16,13 +16,5 @@
 # limitations under the License.
 #
 
-require "chef/chef_fs/file_system/operation_failed_error"
-
-class Chef
-  module ChefFS
-    module FileSystem
-      class AlreadyExistsError < OperationFailedError
-      end
-    end
-  end
-end
+require "chef/chef_fs/file_system/exceptions"
+Chef.log_deprecation "Individual ChefFS error files are deprecated. Please require 'chef/chef_fs/file_system/exceptions' rather than 'chef/chef_fs/file_system/#{File.basename(__FILE__, ".rb")}'."
