+++ lib/chef/chef_fs/file_system/chef_server/acl_dir.rb
@@ -18,7 +18,7 @@
 
 require "chef/chef_fs/file_system/base_fs_dir"
 require "chef/chef_fs/file_system/chef_server/acl_entry"
-require "chef/chef_fs/file_system/operation_not_allowed_error"
+require "chef/chef_fs/file_system/exceptions"
 
 class Chef
   module ChefFS
