+++ lib/chef/chef_fs/file_system/chef_server/acl_entry.rb
@@ -17,9 +17,7 @@
 #
 
 require "chef/chef_fs/file_system/chef_server/rest_list_entry"
-require "chef/chef_fs/file_system/not_found_error"
-require "chef/chef_fs/file_system/operation_not_allowed_error"
-require "chef/chef_fs/file_system/operation_failed_error"
+require "chef/chef_fs/file_system/exceptions"
 
 class Chef
   module ChefFS
