+++ lib/chef/chef_fs/file_system/chef_server/data_bag_dir.rb
@@ -17,8 +17,7 @@
 #
 
 require "chef/chef_fs/file_system/chef_server/rest_list_dir"
-require "chef/chef_fs/file_system/not_found_error"
-require "chef/chef_fs/file_system/must_delete_recursively_error"
+require "chef/chef_fs/file_system/exceptions"
 require "chef/chef_fs/data_handler/data_bag_item_data_handler"
 
 class Chef
