+++ lib/chef/chef_fs/file_system/chef_server/environments_dir.rb
@@ -18,8 +18,7 @@
 
 require "chef/chef_fs/file_system/base_fs_dir"
 require "chef/chef_fs/file_system/chef_server/rest_list_entry"
-require "chef/chef_fs/file_system/not_found_error"
-require "chef/chef_fs/file_system/default_environment_cannot_be_modified_error"
+require "chef/chef_fs/file_system/exceptions"
 
 class Chef
   module ChefFS
