+++ lib/chef/chef_fs/file_system/chef_server/policy_group_entry.rb
@@ -16,9 +16,7 @@
 # limitations under the License.
 #
 
-require "chef/chef_fs/file_system/already_exists_error"
-require "chef/chef_fs/file_system/not_found_error"
-require "chef/chef_fs/file_system/operation_failed_error"
+require "chef/chef_fs/file_system/exceptions"
 
 class Chef
   module ChefFS
