+++ lib/chef/chef_fs/file_system/nonexistent_fs_object.rb
@@ -17,7 +17,7 @@
 #
 
 require "chef/chef_fs/file_system/base_fs_object"
-require "chef/chef_fs/file_system/not_found_error"
+require "chef/chef_fs/file_system/exceptions"
 
 class Chef
   module ChefFS
