+++ lib/chef/chef_fs/file_system/operation_not_allowed_error.rb
@@ -16,31 +16,5 @@
 # limitations under the License.
 #
 
-require "chef/chef_fs/file_system/file_system_error"
-
-class Chef
-  module ChefFS
-    module FileSystem
-      class OperationNotAllowedError < FileSystemError
-        def initialize(operation, entry, cause = nil, reason = nil)
-          reason ||=
-            case operation
-            when :delete
-              "cannot be deleted"
-            when :write
-              "cannot be updated"
-            when :create_child
-              "cannot have a child created under it"
-            when :read
-              "cannot be read"
-            end
-          super(entry, cause, reason)
-          @operation = operation
-        end
-
-        attr_reader :operation
-        attr_reader :entry
-      end
-    end
-  end
-end
+require "chef/chef_fs/file_system/exceptions"
+Chef.log_deprecation "Individual ChefFS error files are deprecated. Please require 'chef/chef_fs/file_system/exceptions' rather than 'chef/chef_fs/file_system/#{File.basename(__FILE__, ".rb")}'."
