+++ lib/chef/chef_fs/file_system/repository/chef_repository_file_system_cookbook_dir.rb
@@ -19,7 +19,7 @@
 require "chef/chef_fs/file_system/repository/chef_repository_file_system_cookbook_entry"
 require "chef/chef_fs/file_system/chef_server/cookbook_dir"
 require "chef/chef_fs/file_system/chef_server/versioned_cookbook_dir"
-require "chef/chef_fs/file_system/not_found_error"
+require "chef/chef_fs/file_system/exceptions"
 require "chef/cookbook/cookbook_version_loader"
 
 class Chef
