+++ lib/chef/chef_fs/file_system/repository/file_system_entry.rb
@@ -18,9 +18,7 @@
 
 require "chef/chef_fs/file_system/base_fs_dir"
 require "chef/chef_fs/file_system/chef_server/rest_list_dir"
-require "chef/chef_fs/file_system/already_exists_error"
-require "chef/chef_fs/file_system/must_delete_recursively_error"
-require "chef/chef_fs/file_system/not_found_error"
+require "chef/chef_fs/file_system/exceptions"
 require "chef/chef_fs/path_utils"
 require "fileutils"
