+++ lib/chef/knife/show.rb
@@ -9,7 +9,7 @@ class Chef
 
       deps do
         require "chef/chef_fs/file_system"
-        require "chef/chef_fs/file_system/not_found_error"
+        require "chef/chef_fs/file_system/exceptions"
       end
 
       option :local,
